Implement authentication and public info services with initial routes
