+++ server/source/Controller/Auth/auth.controller.ts
@@ -1,0 +1,1 @@
+console.log("Auth controller initialized");
