+++ server/source/Services/Auth/Login.service.ts
@@ -1,0 +1,1 @@
+console.log("Login service initialized");
