+++ server/source/Services/Public/Health.service.ts
@@ -1,0 +1,15 @@
+
+/**
+ * Service class to check the health of the application.
+ * 
+ * @export
+ * @class HealthService
+ */
+export default class HealthService {
+  static async checkHealth(): Promise<any> {
+    return {
+      status: "ok",
+      timestamp: new Date().toISOString(),
+    };
+  }
+}
